configuration: allow some trimming options to be optional
There is a balance to be struck when trimming empty lines. Trimming the
empty lines at the bottom and the repeated empty lines within a file
could be seen as too much noise for a diff. Too often if either of the
previous mentioned cases are violated it is more than just one or two
instances within the file and really starts to detract from the
important parts of the diff. Instead, these features should be
considered options, easily enabled or disabled, either at runtime or
startup.